News_analys_bot/requirements.txt
python-telegram-bot==13.7
openai==0.27.8
pandas==2.0.3
python-dotenv==1.0.0
